# Optimize langfst performance: reduce test time from 24s to 10s