## Summary

This PR addresses the performance issue in the `parserule` crate where the `test_build_realistic_lang_fst1` test was taking approximately 22-24 seconds instead of the expected 0.5 seconds.

## Performance Improvement

- **Before**: ~24.53 seconds
- **After**: ~10.05 seconds  
- **Improvement**: >50% reduction in execution time

## Changes Made

### 1. Removed Excessive Debug Output
- Removed debug prints from `mapparse.rs` that were printing parsed rules and symbols
- Removed debug prints from `langfst.rs` that were printing AST structures

### 2. Optimized Mapping FST Construction
- **Before**: Created individual FSTs for each mapping rule and used expensive union operations
- **After**: Build the mapping FST directly by adding all paths to the same FST structure
- This avoids the exponential growth in FST size that was occurring with repeated union operations

### 3. Reduced Redundant Optimizations
- **Before**: FSTs were optimized multiple times (after each union operation and individually)
- **After**: Only optimize the mapping FST once, which is the most complex and benefits most from optimization
- Skip optimization of preprocessing and postprocessing FSTs which are typically simpler

### 4. Code Cleanup
- Removed unused imports (`union::union`, `DeterminizeConfig`, `DeterminizeType`, `tr_sum`)
- Cleaned up commented code

## Testing

- All 94 existing tests continue to pass
- The optimized `test_build_realistic_lang_fst1` consistently runs in ~10 seconds across multiple runs
- No functional changes - all FST operations produce the same results

## Technical Details

The main bottleneck was in the `compile_mapping_fst` function where:
1. Each mapping rule created a separate transducer FST
2. Each transducer was individually optimized
3. Each transducer was unioned with the main FST
4. The main FST was re-optimized after each union

This created an O(n²) complexity issue where each union operation became progressively more expensive as the FST grew larger.

The new approach builds the FST directly by sharing common prefixes and states, resulting in a much more efficient construction process.